add changes on usercontex and moved files to shared-component
