+++ src/shared-component/Button/index.jsx
@@ -1,6 +1,5 @@
 import React from "react";
-import "../../App.css"; // Ensure to import the CSS file
-
+import "./styles.css";
 export const Button = ({ children, onClick }) => {
   return (
     <button className="shared-button" onClick={onClick}>
