+++ src/shared-component/Button/styles.css
@@ -1,0 +1,16 @@
+.shared-button {
+  margin-bottom: 10px;
+  padding: 10px 20px;
+  background-color: #FF6347;
+  color: white;
+  border: none;
+  border-radius: 8px;
+  cursor: pointer;
+  font-size: 16px;
+  transition: background-color 0.3 ease;
+  width: 100%;
+}
+
+.shared-button:hover {
+  background-color: #FF4500;
+}
