+++ src/shared-component/Container/styles.css
@@ -1,0 +1,12 @@
+.shared-container {
+  max-width: 400px;
+  margin: 50px auto;
+  padding: 20px;
+  border: 1px solid #ccc;
+  border-radius: 8px;
+  background-color: #f9f9f9;
+  box-shadow: 0 2px 4px #0000001a;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
