+++ src/shared-component/Input/styles.css
@@ -1,0 +1,12 @@
+.shared-input {
+  width: 100%;
+  padding: 10px;
+  margin: 10px 0;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+}
+
+.shared-h1 {
+  text-align: center;
+margin-bottom: 20px;
+}
